requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-cors-headers==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
mysqlclient==1.3.12
pytz==2018.3
requests==2.18.4
unicodecsv==0.14.1
urllib3==1.22
